tests/wpt/web-platform-tests/css/css-box/parsing/max-height-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: parsing max-height with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-max-height">
<meta name="assert" content="max-height supports the full grammar '[ [<length> | <percentage>] && [border-box | content-box]? ] | available | min-content | max-content | fit-content | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("max-height", "none");

test_valid_value("max-height", "10px");
test_valid_value("max-height", "20%");
test_valid_value("max-height", "calc(2em + 3ex)");

test_valid_value("max-height", "min-content");
test_valid_value("max-height", "max-content");
test_valid_value("max-height", "fit-content");

// The following are not yet supported by browsers:
// test_valid_value("max-height", "available");
// test_valid_value("max-height", "10px border-box");
// test_valid_value("max-height", "content-box 20%", "20% content-box");
</script>
</body>
</html>
